// dfp-piao-fe/src/app/shared/models/enums/cod-tipologia-allegato.enum.ts
export enum CodTipologiaAllegatoEnum {
  LOGO_ANAGRAFICA = 'LOGO_ANAGRAFICA',
  ALLEGATO_SEZIONE = 'ALLEGATO_SEZIONE',
  ANALISI_CONTESTO = 'ANALISI_CONTESTO',
  VALORE_PUBBLICO = 'VALORE_PUBBLICO',
  IMMAGINE_SEZIONE_21_1 = 'IMMAGINE_SEZIONE_21_1',
  IMMAGINE_SEZIONE_21_2 = 'IMMAGINE_SEZIONE_21_2',
  IMMAGINE_SEZIONE_31 = 'IMMAGINE_SEZIONE_31',
  MANCATA_COMPILAZIONE = 'MANCATA_COMPILAZIONE',
  PIAO = 'PIAO',
  ULTERIORI_ALLEGATI_PIAO = 'ULTERIORI_ALLEGATI_PIAO',
  DOCUMENTAZIONE_MONITORAGGIO = 'DOCUMENTAZIONE_MONITORAGGIO',
  RICHIESTA_APPROVAZIONE = 'RICHIESTA_APPROVAZIONE',
  APPROVAZIONE_PUBBLICAZIONE = 'APPROVAZIONE_PUBBLICAZIONE',
  HELP_DESK = 'HELP_DESK',
  PIAO_PDF_GENERATO = 'PIAO_PDF_GENERATO',
  AVVISI_DFP = 'AVVISI_DFP',
}